refactor: extract series feature operations
